fix: move stuff around
